app/requirements.txt
aiogram==3.12.00
sqlalchemy==2.0.32
apscheduler==3.10.4
aiohttp==3.10.4
python-dotenv==1.0.1